+++ src/directiveClassSuffixRule.ts
@@ -6,10 +6,10 @@ import {Ng2Walker} from './angular/ng2Walker';
 import {DirectiveMetadata} from './angular/metadata';
 
 export class Rule extends Lint.Rules.AbstractRule {
-  static FAILURE:string = 'The name of the class %s should end with the suffix Directive ($$02-03$$)';
+  static FAILURE:string = 'The name of the class %s should end with the suffix %s ($$02-03$$)';
 
-  static validate(className: string):boolean {
-    return /.*Directive/.test(className);
+  static validate(className: string, suffix: string):boolean {
+    return className.endsWith(suffix);
   }
 
   public apply(sourceFile:ts.SourceFile):Lint.RuleFailure[] {
@@ -23,12 +23,13 @@ export class ClassMetadataWalker extends Ng2Walker {
   visitNg2Directive(meta: DirectiveMetadata) {
     let name = meta.controller.name;
     let className:string = name.text;
-    if (!Rule.validate(className)) {
+    const suffix = this.getOptions()[0] || 'Directive';
+    if (!Rule.validate(className, suffix)) {
       this.addFailure(
         this.createFailure(
           name.getStart(),
           name.getWidth(),
-          sprintf.apply(this, [Rule.FAILURE, className])));
+          sprintf.apply(this, [Rule.FAILURE, className, suffix])));
     }
   }
 }